% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_stats_column}
\alias{summary_stats_column}
\title{Summary Statistics for a Numeric Column in a Dataframe}
\usage{
summary_stats_column(data, col)
}
\arguments{
\item{data}{The data.frame to be inputted. This is named \code{data} because it describes what the first argument should be — a data.frame.}

\item{col}{The numerical column to be used for calculating summary statistics. This is named \code{col} because it describes what the second argument should be — a column.}
}
\value{
A data.frame of summary statistics (mean, median, max, min, standard deviation, and count), as well as a histogram.
}
\description{
This function takes in a dataframe and a numerical column in that dataframe and outputs a data.frame of summary statistics (mean, median, max, min, standard deviation, and count), as well as a histogram of that column's distribution with a vertical line representing the mean.
}
\examples{
summary_stats_column(mtcars,hp)
summary_stats_column(mtcars, mpg)
}
